# discord-bot/backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Discord Automation Cloud SaaS API",
    "version": "1.0.0",
    "description": "Production REST API for multi-tenant Discord bot automation platform, plugin settings, AI assistants, tickets, visual workflows, and billing."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local API Gateway"
    }
  ],
  "paths": {
    "/guilds": {
      "get": {
        "summary": "List all active Discord guilds where bot exists",
        "responses": {
          "200": {
            "description": "Successful response with array of guilds"
          }
        }
      }
    },
    "/guilds/{guildId}/plugins": {
      "get": {
        "summary": "Get all plugin configurations for a guild",
        "parameters": [
          { "name": "guildId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Plugins list with toggle states" }
        }
      }
    },
    "/guilds/{guildId}/plugins/{pluginKey}/toggle": {
      "post": {
        "summary": "Enable or disable a plugin in real time",
        "parameters": [
          { "name": "guildId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "pluginKey", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Plugin toggle state updated" }
        }
      }
    },
    "/ai/chat": {
      "post": {
        "summary": "Execute multi-model AI assistant prompt with RAG context",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" },
                  "model": { "type": "string", "enum": ["gpt-4o", "gemini-1.5-pro", "claude-3-sonnet", "deepseek-r1"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "AI reply with token metadata" }
        }
      }
    },
    "/automations": {
      "post": {
        "summary": "Create or update visual n8n-style automation rule",
        "responses": {
          "200": { "description": "Automation rule created" }
        }
      }
    }
  }
}